Refactor existing integration tests
