+++ pimi-connect-api/pimi-connect-api.Tests/IntegrationTests/Base/IntegrationTestsBase.cs
@@ -2,11 +2,10 @@ using AutoMapper;
 using Microsoft.Extensions.Configuration;
 using pimi_connect_api.UnitTests.Shared;
 using pimi_connect_app.Data.AppDbContext;
-using pimi_connect_app.Data.MappingProfiles;
 
-namespace pimi_connect_api.UnitTests.Base;
+namespace pimi_connect_api.Tests.IntegrationTests.Base;
 
-public abstract class ControllerTestsBase<TDtoType>  where TDtoType : class
+public abstract class IntegrationTestsBase<TDtoType>  where TDtoType : class
 {
     #region Properties
     protected AppDbContext TestDbContext { get; private set; }
@@ -19,7 +18,7 @@ public abstract class ControllerTestsBase<TDtoType>  where TDtoType : class
     protected Guid NotExistingId { get; private set; }
     #endregion
     
-    protected ControllerTestsBase()
+    protected IntegrationTestsBase()
     {
         SetupConfiguration();
         SetupSettings();
@@ -85,8 +84,6 @@ public abstract class ControllerTestsBase<TDtoType>  where TDtoType : class
             .Build();
 
         Configuration = configuration;
-
-        
     }
 
     private void SetupSettings()
@@ -103,12 +100,7 @@ public abstract class ControllerTestsBase<TDtoType>  where TDtoType : class
     
     private void SetupMapper()
     {
-        var mapperConfig = new MapperConfiguration(cfg => 
-        {
-            cfg.AddProfile(new MappingProfile());
-        });
-
-        Mapper = new Mapper(mapperConfig);
+        Mapper = Utils.CreateMapper();
     }
     
     private void SetupTestDbContext()
@@ -133,12 +125,7 @@ public abstract class ControllerTestsBase<TDtoType>  where TDtoType : class
     
     private void SetExistingIds()
     {
-        ExistingIds = new List<Guid>();
-        
-        for (var i = 0; i < Settings.EntitiesCount; i++)
-        {
-            ExistingIds.Add(Guid.NewGuid());
-        }
+        ExistingIds = Utils.CreateGuids(Settings.EntitiesCount);
     }
 
     private void SetExistingId()
